Introduced STM and bumped the version of xio-http to 0.0.4

+++ src/main/scala/com/kabasoft/iws/api/JournalRoutes.scala
@@ -1,11 +1,9 @@
 package com.kabasoft.iws.api
 
 import com.kabasoft.iws.api.Protocol._
-import com.kabasoft.iws.domain._
 import com.kabasoft.iws.repository._
-import zio._
 import zio.http._
-import zio.http.model.{Method, Status}
+import zio.http.model.Method
 import zio.json._
 
 object JournalRoutes {
@@ -19,7 +17,7 @@ object JournalRoutes {
           .map(ch => Response.json(ch.toJson))
 
       case Method.GET -> !! / "jou" / id   =>
-        JournalRepository.getBy(id, "1000").either.map {
+        JournalRepository.getBy(id.toLong, "1000").either.map {
           case Right(o) => Response.json(o.toJson)
           case Left(e)  => Response.text(e.getMessage() + "ID" + id)
         }
